+++ src/js/desktop/appInit.js
@@ -179,19 +179,22 @@ chrome.storage.local.get(['vkAccessToken', 'vkUserID'], function (userAuthData) 
     })
 
     domManipulations.hideSpinner()
+    domManipulations.showToast({
+      'innerText': (chrome.i18n.getMessage('initialToast') ||
+        'Інформація для підключення віддаленого керування знаходиться на сторінці налаштувань'),
+      'duration': 10000
+    })
   }).catch(function (err) {
     domManipulations.hideSpinner()
+    var errMsg
+    if (+err.error_code === 5) {
+      errMsg = chrome.i18n.getMessage('userAuthorizationFailed') || 'Сталася помилка :( Будь ласка, вийдіть і залогіньтесь знову.'
+    } else {
+      errMsg = chrome.i18n.getMessage('errorOccurred') || 'Сталася помилка'
+    }
     domManipulations.showToast({
-      'innerText': chrome.i18n.getMessage('errorOccurred') || 'Сталася помилка'
-    })
-    connectedSockets.forEach(function (socket) {
-      sendDomManipulationsMessage({
-        'socket': socket,
-        'function': 'showToast',
-        'args': {
-          'innerText': chrome.i18n.getMessage('errorOccurred') || 'Сталася помилка'
-        }
-      })
+      'innerText': errMsg,
+      'duration': 9999999
     })
     console.error(err)
   })